Use RES_GET/PUT in place of NS_GET/PUT. These macros automatically increment the buf pointer as it reads or stores data. Also use u_int16_t class, u_int16_t type in the res_val_nmkquery prototype

+++ dnssec-tools/validator/libsres/res_mkquery.h
@@ -26,7 +26,7 @@
 int
                 res_val_nmkquery(struct name_server *pref_ns, int op,   /* opcode of query */
                                  const char *dname,     /* domain name */
-                                 int class, int type,   /* class and type of query */
+                                 u_int16_t class, u_int16_t type,   /* class and type of query */
                                  const u_char * data,   /* resource record data */
                                  int datalen,   /* length of data */
                                  const u_char * newrr_in,       /* new rr for modify or append */
